% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsem.R
\name{.lavaan_vars}
\alias{.lavaan_vars}
\title{Internal function to extract variable names for each tested model
This is used to find largest umber of folds K}
\usage{
.lavaan_vars(X, data)
}
\arguments{
\item{X}{lavaan model}

\item{data}{Data}
}
\value{
list of variable names
}
\description{
Internal function to extract variable names for each tested model
This is used to find largest umber of folds K
}
\author{
philippe
}
\keyword{internal}
