% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{fd}
\alias{fd}
\title{Frobenius Matrix Discrepancy}
\usage{
fd(implied_Sigma, test_S)
}
\arguments{
\item{implied_Sigma}{Model implied covariances matrix from training set}

\item{test_S}{Sample covariance matrix from test set}
}
\value{
FD discrepancy
}
\description{
Frobenius Distance as described in \insertCite{Biscay1997}{cvsem} or \insertCite{Amendola2015}{cvsem}.
}
\references{
\insertAllCited()
}
