% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{gls}
\alias{gls}
\title{Generalized Least Squares Discrepancy Function}
\usage{
gls(implied_sigma, test_S)
}
\arguments{
\item{implied_sigma}{Model implied covariances matrix from training set}

\item{test_S}{Sample covariance matrix from test set}
}
\value{
GLS discrepancy
}
\description{
Generalized Least Squares (GLS) Discrepancy as defined in \insertCite{Cudeck1983}{cvsem}.
}
\references{
\insertAllCited()
}
