% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{KL_divergence}
\alias{KL_divergence}
\title{Compute KL-Divergence on two covariance matrices. KL-Divergence corresponds to the
Maximum Wishart Likelihood (MWL) discrepancy described in \insertCite{Cudeck1983}{cvsem}.}
\usage{
KL_divergence(implied_sigma, test_S)
}
\arguments{
\item{implied_sigma}{Model implied covariances matrix from training set}

\item{test_S}{Sample covariance matrix from test set}
}
\value{
KL-Divergence index
}
\description{
Compute KL-Divergence on two covariance matrices. KL-Divergence corresponds to the
Maximum Wishart Likelihood (MWL) discrepancy described in \insertCite{Cudeck1983}{cvsem}.
}
\references{
\insertAllCited()
}
