% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cvsem}
\alias{print.cvsem}
\title{Print cvsem object}
\usage{
\method{print}{cvsem}(x, digits = 2, ...)
}
\arguments{
\item{x}{cvsem object}

\item{digits}{Round to (default 2) digits.}

\item{...}{not used}
}
\value{
Formatted cvsem object
}
\description{
Return the ordered list of models where the model with the smallest discrepancy metric is listed first.
}
