% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsem.R
\name{cvsem}
\alias{cvsem}
\title{Do model comparison on SEM models using cross-validation}
\usage{
cvsem(x, Models, k = 5)
}
\arguments{
\item{x}{Data}

\item{Models}{A list of specified models in lavaan syntax}

\item{k}{The number of folds. Default is 5.}
}
\value{
A list with the prediction error for each model.
}
\description{
Do model comparison on SEM models using cross-validation
}
\examples{
data <- c(1, 2, 3)
Models <- "xx"
cvsem(x = data, Models = Models)
}
