% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsem.R
\name{cvsem}
\alias{cvsem}
\title{Do model comparison on SEM models using cross-validation}
\usage{
cvsem(
  x,
  Models,
  distanceMetric = "KL-Divergence",
  k = 5,
  lavaanFunction = "sem",
  echo = TRUE,
  ...
)
}
\arguments{
\item{x}{Data}

\item{Models}{A collection of models, specified in lavaan syntax. Provide Models with the \code{cvgather()} function.}

\item{distanceMetric}{Specify which distance metric to use. Default is KL Divergence. Other option is the Maximum Wishart Likelihood (MWL)}

\item{k}{The number of folds. Default is 5.}

\item{lavaanFunction}{Specify which lavaan function to use. Default is "sem". Other options are "lavaan" and "cfa"}

\item{echo}{Provide feeback on progess to user, defaults to \code{TRUE}. Set to \code{FALSE} to supppress.}

\item{...}{Not used}
}
\value{
A list with the prediction error for each model.
}
\description{
Do model comparison on SEM models using cross-validation
}
\examples{

example_data <- lavaan::HolzingerSwineford1939
colnames(example_data) <- c("id", "sex", "ageyr", "agemo", 'school', "grade",
"visualPerception", "cubes", "lozenges", "comprehension",
"sentenceCompletion", "wordMeaning", "speededAddition",
"speededCounting", "speededDiscrimination")

model2 <- 'comprehension ~ meaning

           ## Add some latent variables:
        meaning =~ wordMeaning + sentenceCompletion
        speed =~ speededAddition + speededDiscrimination + speededCounting
        speed ~~ meaning'

model3 <- 'comprehension ~ wordMeaning + speededAddition'

model_list <- cvgather(model1, model2)

}
